anvil: modified a dependency (revm-handler) that calculated gas usage wrong
The dependency is responsible for the instrinsic costs of a EIP-7702 transaction, and assumes the wrong costs. Because anvil's database always fetches an account, an empty account is not loaded 'correctly' in terms of revm-handler and therefore, the wrong parameter of EIP-7702 is selected and a gas refund is applied that should not be there. There has been discussion about this phenomena on github: https://github.com/bluealloy/revm/issues/2938